Close the source on fclose
This so far was not a problem, because zip_fopen was returning a file
with toplevel source with refcount of 1. When we were closing the
file, the source was also destroyed and, as a part of the destruction,
the source was also closed.

Now, with the reopen functionality, we may return an already existing
source, so its refcount is going to be higher than 1. In this
situation, the source was not destroyed, and was not closed. This
could result in an empty file being written into the archive, if the
file was read whole earlier.